Start renaming countsserver module to ondemand to reflect expanded role.
