+++ spatialprofilingtoolbox/ondemand/counts_service_client.py
