+++ test/ondemand/module_tests/check_binary_expressions_equivalent.sh
