+++ test/ondemand/module_tests/expected_binary_expression_data_cohoused/centroids.pickle
